completely removed v1 api and its usages
