+++ src/helper/utils.ts
@@ -1,6 +1,5 @@
-import { UserCookies } from "types/user.types";
+import { UserCookies } from "types/user-v2.types";
 import { UserCookies as UserCookiesV2 } from "types/user-v2.types";
-import { useUserStore } from "stores/userStore_v1";
 
 // export const getCookiesAsString = (cookies: UserCookies) => cookies.fhwn + "; " + cookies.session;
 // export const getCookiesAsStringV2 = (cookies: UserCookiesV2) =>
